+++ CHANGELOG.md
@@ -13,6 +13,7 @@ Unreleased
 * Changed 'linsolver' param value of nr to 'default'.
 * INLP improvements: separate primal and dual steps, better var initializations.
 * Fixed IPOPT windows dll handling.
+* Seperated feasibility and optimality tolerances in inlp.
 
 Version 1.1.5
 -------------